+++ scripts/run-terraform-locally.sh
@@ -2,7 +2,7 @@
 
 #
 # *** Script Syntax ***
-# scripts/run-terraform-locally.sh --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>
+# scripts/run-terraform-locally.sh --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>
 #
 #
 
@@ -12,7 +12,7 @@ then
     echo
     echo "(Error Message 001)  You did not include all four arguments in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
@@ -40,6 +40,15 @@ do
         *"--action=destroy"*)
             action_argument_supplied=true
             create_action=false;;
+        *"--snowflake_login_url="*)
+            arg_length=22
+            snowflake_login_url=${arg:$arg_length:$(expr ${#arg} - $arg_length)};;
+        *"--snowflake_username="*)
+            arg_length=21
+            snowflake_username=${arg:$arg_length:$(expr ${#arg} - $arg_length)};;
+        *"--snowflake_password="*)
+            arg_length=21
+            snowflake_password=${arg:$arg_length:$(expr ${#arg} - $arg_length)};;
     esac
     let "arg_count+=1"
 done
@@ -50,7 +59,7 @@ then
     echo
     echo "(Error Message 002)  You did not include the proper use of the --environment=<ENVIRONMENT_NAME> argument in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
@@ -61,7 +70,7 @@ then
     echo
     echo "(Error Message 003)  You did not include the proper use of the --profile=<AWS_SSO_SSO_PROFILE_NAME> argument in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
@@ -72,7 +81,7 @@ then
     echo
     echo "(Error Message 004)  You did not include the proper use of the --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> argument in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
@@ -83,7 +92,7 @@ then
     echo
     echo "(Error Message 005)  You did not include the proper use of the --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> argument in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
@@ -92,13 +101,45 @@ fi
 if [ "$action_argument_supplied" = false ]
 then
     echo
-    echo "(Error Message 002)  You did not include the proper use of the --action=<create | destroy> argument in the call."
+    echo "(Error Message 006)  You did not include the proper use of the --action=<create | destroy> argument in the call."
     echo
-    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy>"
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
     echo
     exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
 fi
 
+# Check required --snowflake_login_url argument was supplied
+if [ -z $snowflake_login_url ]
+then
+    echo
+    echo "(Error Message 007)  You did not include the proper use of the --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> argument in the call."
+    echo
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
+    echo
+    exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
+fi
+
+# Check required --snowflake_username argument was supplied
+if [ -z $snowflake_username ]
+then
+    echo
+    echo "(Error Message 008)  You did not include the proper use of the --snowflake_username=<SNOWFLAKE_USERNAME> argument in the call."
+    echo
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
+    echo
+    exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
+fi
+
+# Check required --snowflake_password argument was supplied
+if [ -z $snowflake_password ]
+then
+    echo
+    echo "(Error Message 009)  You did not include the proper use of the --snowflake_password=<SNOWFLAKE_PASSWORD> argument in the call."
+    echo
+    echo "Usage:  Require all four arguments ---> `basename $0` --environment=<ENVIRONMENT_NAME> --profile=<SSO_PROFILE_NAME> --confluent_cloud_api_key=<CONFLUENT_CLOUD_API_KEY> --confluent_cloud_api_secret=<CONFLUENT_CLOUD_API_SECRETS> --action=<create | destroy> --snowflake_login_url=<SNOWFLAKE_LOGIN_URL> --snowflake_username=<SNOWFLAKE_USERNAME> --snowflake_password=<SNOWFLAKE_PASSWORD>"
+    echo
+    exit 85 # Common GNU/Linux Exit Code for 'Interrupted system call should be restarted'
+fi
 
 # Get the SSO AWS_ACCESS_KEY_ID, AWS_ACCESS_SECRET_KEY, AWS_SESSION_TOKEN, and AWS_REGION, and
 # set them as an environmental variables
@@ -116,6 +157,9 @@ printf "confluent_cloud_api_key=\"${confluent_cloud_api_key}\"\
 \naws_access_key_id=\"${AWS_ACCESS_KEY_ID}\"\
 \naws_secret_access_key=\"${AWS_SECRET_ACCESS_KEY}\"\
 \naws_session_token=\"${AWS_SESSION_TOKEN}\"\
+\nsnowflake_login_url=\"${snowflake_login_url}\"\
+\nsnowflake_username=\"${snowflake_username}\"\
+\nsnowflake_password=\"${snowflake_password}\"\
 \nnumber_of_api_keys_to_retain = 2\
 \nday_count=30\
 \nauto_offset_reset=\"earliest\"" > terraform.tfvars